+++ h2-web-console/h2webconsole-service/src/main/java/org/lealone/examples/h2webconsole/service/QueryServiceImpl.java
@@ -50,8 +50,6 @@ import org.lealone.orm.json.JsonObject;
 
 public class QueryServiceImpl extends ServiceImpl implements QueryService {
 
-    ServiceSession session;
-
     private Profiler profiler;
 
     /**
